feat: Add GitHub Actions CI workflow
This commit introduces a new Continuous Integration (CI) workflow using
GitHub Actions. The workflow is triggered on pull requests and pushes
to the main branch or any tags. The jobs include setting up QEMU and
Docker Buildx, logging into GHCR, and building and pushing Docker
images to GHCR. The Docker image tags include the latest tag and a
tag with the commit SHA and date.